added lodepng for image saving stuff

+++ include/CMU462/spectrum.h
@@ -14,9 +14,9 @@ namespace CMU462 {
  */
 class Spectrum {
  public:
-  double r;  ///< intensity of red spectrum
-  double g;  ///< intensity of green spectrum
-  double b;  ///< intensity of blue spectrum
+  float r;  ///< intensity of red spectrum
+  float g;  ///< intensity of green spectrum
+  float b;  ///< intensity of blue spectrum
 
   /**
    * Parameterized Constructor.
@@ -25,7 +25,7 @@ class Spectrum {
    * \param g Intensity of the green spectrum
    * \param b Intensity of the blue spectrum
    */
-  Spectrum(double r = 0, double g = 0, double b = 0) : r(r), g(g), b(b) {}
+  Spectrum(float r = 0, float g = 0, float b = 0) : r(r), g(g), b(b) {}
 
   /**
    * Constructor.
@@ -58,11 +58,11 @@ class Spectrum {
     return *this;
   }
 
-  inline Spectrum operator*(double s) const {
+  inline Spectrum operator*(float s) const {
     return Spectrum(r * s, g * s, b * s);
   }
 
-  inline Spectrum &operator*=(double s) {
+  inline Spectrum &operator*=(float s) {
     r *= s;
     g *= s;
     b *= s;
@@ -86,7 +86,7 @@ class Spectrum {
 };  // class Spectrum
 
 // Commutable scalar multiplication
-inline Spectrum operator*(double s, const Spectrum &c) { return c * s; }
+inline Spectrum operator*(float s, const Spectrum &c) { return c * s; }
 
 // Prints components
 std::ostream &operator<<(std::ostream &os, const Spectrum &c);
